# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Datespark API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/match/get-potential-matches": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Get potential matches for logged user",
                "operationId": "ce202e8e1aa8b4e1e46a65be01772a4d",
                "responses": {
                    "200": {
                        "description": "Return list of user which is potential match",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/match/get-matches": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Get matches for logged user",
                "operationId": "3b1264f941e5b6f429ce5def89256ea6",
                "responses": {
                    "200": {
                        "description": "Return list of user which is match",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/match/accept-match": {
            "post": {
                "tags": [
                    "Match"
                ],
                "summary": "Create or accept match of a given user for logged user",
                "operationId": "2a3c690fafc58429652bf90d6cb2da78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idOfUserUserWantToMatch"
                                ],
                                "properties": {
                                    "idOfUserUserWantToMatch": {
                                        "description": "Id of user which logged user want to accept",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return message with success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/match/reject-match": {
            "post": {
                "tags": [
                    "Match"
                ],
                "summary": "Create or reject match of a given user for logged user",
                "operationId": "66a27b569d0f2da28e586a96d671a5bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idOfUserUserWantToMatch"
                                ],
                                "properties": {
                                    "idOfUserUserWantToMatch": {
                                        "description": "Id of user which logged user want to reject",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return message with success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Invalid body"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "description": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "surname": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "email": {
                        "type": "string",
                        "example": "jan@example.com"
                    },
                    "status": {
                        "description": "Either ACTIVE or DURING_REGISTER",
                        "type": "string",
                        "example": "ACTIVE"
                    },
                    "gender": {
                        "description": "Either MALE or FEMALE",
                        "type": "string",
                        "example": "MALE"
                    },
                    "roles": {
                        "description": "Array of roles of user in system",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "USER",
                                "ADMIN"
                            ]
                        },
                        "example": [
                            "USER"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}